Add logic for set mutations from DirectedEdges.
