+++ worker/worker.go
@@ -2,13 +2,11 @@ package worker
 
 import (
 	"flag"
-	"fmt"
 	"io"
 	"net"
 	"net/rpc"
 
 	"github.com/dgraph-io/dgraph/conn"
-	"github.com/dgraph-io/dgraph/posting"
 	"github.com/dgraph-io/dgraph/store"
 	"github.com/dgraph-io/dgraph/task"
 	"github.com/dgraph-io/dgraph/x"
@@ -64,87 +62,6 @@ func Connect(workerList []string) {
 	glog.Info("Server started. Clients connected.")
 }
 
-func ProcessTaskOverNetwork(qu []byte) (result []byte, rerr error) {
-	uo := flatbuffers.GetUOffsetT(qu)
-	q := new(task.Query)
-	q.Init(qu, uo)
-
-	attr := string(q.Attr())
-	idx := farm.Fingerprint64([]byte(attr)) % numInstances
-
-	var runHere bool
-	if attr == "_xid_" || attr == "_uid_" {
-		idx = 0
-		runHere = (instanceIdx == 0)
-	} else {
-		runHere = (instanceIdx == idx)
-	}
-
-	if runHere {
-		return ProcessTask(qu)
-	}
-
-	pool := pools[idx]
-	addr := pool.Addr
-	query := new(conn.Query)
-	query.Data = qu
-	reply := new(conn.Reply)
-	if err := pool.Call("Worker.ServeTask", query, reply); err != nil {
-		glog.WithField("call", "Worker.ServeTask").Fatal(err)
-	}
-	glog.WithField("reply", string(reply.Data)).WithField("addr", addr).
-		Info("Got reply from server")
-	return reply.Data, nil
-}
-
-func ProcessTask(query []byte) (result []byte, rerr error) {
-	uo := flatbuffers.GetUOffsetT(query)
-	q := new(task.Query)
-	q.Init(query, uo)
-	attr := string(q.Attr())
-
-	b := flatbuffers.NewBuilder(0)
-	voffsets := make([]flatbuffers.UOffsetT, q.UidsLength())
-	uoffsets := make([]flatbuffers.UOffsetT, q.UidsLength())
-
-	for i := 0; i < q.UidsLength(); i++ {
-		uid := q.Uids(i)
-		key := posting.Key(uid, attr)
-		pl := posting.GetOrCreate(key, dataStore)
-
-		var valoffset flatbuffers.UOffsetT
-		if val, err := pl.Value(); err != nil {
-			valoffset = b.CreateByteVector(x.Nilbyte)
-		} else {
-			valoffset = b.CreateByteVector(val)
-		}
-		task.ValueStart(b)
-		task.ValueAddVal(b, valoffset)
-		voffsets[i] = task.ValueEnd(b)
-
-		ulist := pl.GetUids()
-		uoffsets[i] = x.UidlistOffset(b, ulist)
-	}
-	task.ResultStartValuesVector(b, len(voffsets))
-	for i := len(voffsets) - 1; i >= 0; i-- {
-		b.PrependUOffsetT(voffsets[i])
-	}
-	valuesVent := b.EndVector(len(voffsets))
-
-	task.ResultStartUidmatrixVector(b, len(uoffsets))
-	for i := len(uoffsets) - 1; i >= 0; i-- {
-		b.PrependUOffsetT(uoffsets[i])
-	}
-	matrixVent := b.EndVector(len(uoffsets))
-
-	task.ResultStart(b)
-	task.ResultAddValues(b, valuesVent)
-	task.ResultAddUidmatrix(b, matrixVent)
-	rend := task.ResultEnd(b)
-	b.Finish(rend)
-	return b.Bytes[b.Head():], nil
-}
-
 func NewQuery(attr string, uids []uint64) []byte {
 	b := flatbuffers.NewBuilder(0)
 	task.QueryStartUidsVector(b, len(uids))
@@ -181,25 +98,8 @@ func (w *Worker) Mutate(query *conn.Query, reply *conn.Reply) (rerr error) {
 	}
 
 	left := new(Mutations)
-	// For now, assume it's all only Set instructions.
-	for _, edge := range m.Set {
-		if farm.Fingerprint64(
-			[]byte(edge.Attribute))%numInstances != instanceIdx {
-
-			glog.WithField("instanceIdx", instanceIdx).
-				WithField("attr", edge.Attribute).
-				Info("Predicate fingerprint doesn't match instanceIdx")
-			return fmt.Errorf("predicate fingerprint doesn't match this instance.")
-		}
-
-		key := posting.Key(edge.Entity, edge.Attribute)
-		plist := posting.GetOrCreate(key, dataStore)
-		if err := plist.AddMutation(edge, posting.Set); err != nil {
-			left.Set = append(left.Set, edge)
-			glog.WithError(err).WithField("edge", edge).
-				Error("While adding mutation.")
-			continue
-		}
+	if err := mutate(m, left); err != nil {
+		return err
 	}
 	reply.Data, rerr = left.Encode()
 	return
@@ -212,7 +112,7 @@ func (w *Worker) ServeTask(query *conn.Query, reply *conn.Reply) (rerr error) {
 	attr := string(q.Attr())
 
 	if farm.Fingerprint64([]byte(attr))%numInstances == instanceIdx {
-		reply.Data, rerr = ProcessTask(query.Data)
+		reply.Data, rerr = processTask(query.Data)
 	} else {
 		glog.WithField("attribute", attr).
 			WithField("instanceIdx", instanceIdx).
